split ktmetrics
